+++ currency/gambling_config.py
@@ -1,6 +1,6 @@
 # On average, the return is just a smidge over 75%, compliant with Nevada law.
 slots_probabilities: list[float] = [0.005, 0.015, 0.02, 0.03333333333333333, 0.05, 0.1, 0.21, 0.5666666666666667]
-slots_payouts: list[int] = [20, 10, 5, 3, 2, 1, 0.5, 0]
+slots_payouts: list[float] = [20.0, 10.0, 5.0, 3.0, 2.0, 1.0, 0.5, 0.0]
 slots_min_bet: int = 1
 slots_max_bet: int = 10_000
 lottery_ticket_price: int = 5
